Refactor tests for speed and helper methods
